Add mechanism to skip getting descriptions in GetBlockInfo and GetFieldInfo. NOTE: This is broken and will not even start a connection to a PandA!
Issue is that when skipping description in GetFieldInfo, some fields end up generating no extra commands as they have no ENUM fields (e.g. LVDSIN), and hence they finish before commands that do generate additional commands. This returns them into the _responses of the ControlConnection "ahead of time" and so they end up mismatched with the initial commands list in the BlockingClient, and we hit the assert in BlockingClient.send() that checks the commands are aligned with responses.