Beginnings of TABLE support
Includes retrieving the MAX_LENGTH and FIELDS attributes from the table fields, and a new FieldInfo type to hold the returned info.

Includes some (in-progress) refactoring of the values dict so it can hold both str and List[str] items - not clear how best to do this cleanly though.

Includes some neatening up of how block and field names are created/manipulated to go between PandA and EPICS conventions.