Add logging to ioc.py
It's mostly replacing print() statements, and some exceptions as well. Some DEBUG level messages added, although there could/probably should be far more of them for true debugging purposes.

Note that an awful lot of PandA field descriptions, and a lot of field labels, are too long to fit in EPICS so there are a LOT of warnings printed on startup.