Change the clients to make repeated calls to data() function correctly.
Previously the clients kept a permanent connection to the server's data port, which was never closed. This meant that even if we disabled capturing by stopping the task running the data capture, the PandA was still sending data to the port that just sat in the receive buffer forever.

Also when a second call to data() was made, a second "XML FRAMED RAW" instruction was send to the PandA, which is ignored and so never responds as expected.