Move TIME attributes onto base record
Now the MIN and UNITS attributes are present on the base record
(as well as UNITS still having its own record, so it can be set)

Updates to the UNITS record are reflected into the .EGU field.

Updates to the MIN value are placed in the record's .DRVL field

Added tests to confirm setting these values from both PandA and EPICS
works as expected.