Add stub functions for better errors for extras
By providing the functions as hidden, they won't appear in any help
output but will run if directly executed. This'll print a more helpful
error for users who may be following a guide and not realizing they need
to specifically install the extras.

Alternative mechanisms like using `logging` don't work well as it means
it'll either always or never print the message.