Fix delayed data frame updates
During data streaming, the pending stage is removed, so
that clients gets `FrameData` as soon as possible. A
disadvantage is that the client may receive some corrupt
data. The user is able to detect this condition, so
the benefit is thought to outweigh the disadvantage.

When writing HDF5 files via `write_hdf_files`, an
exception will now be raised if an overrun occurs.

Documentation and tests updated accordingly.

Closes #54